V56 modulator block added
with all sub blocks! changed samp_rate into float for all blocks

+++ include/ecss/pll.h
@@ -59,7 +59,7 @@ namespace gr {
         * \param freq_central central value of frequency that PLL can catch.
         * \param bw bandwidth of frequency that PLL can catchl.
        */
-      static sptr make(int samp_rate, int order, int N, double Coeff1_2, double Coeff2_2, double Coeff4_2, double Coeff1_3, double Coeff2_3, double Coeff3_3, float freq_central, float bw);
+      static sptr make(float samp_rate, int order, int N, double Coeff1_2, double Coeff2_2, double Coeff4_2, double Coeff1_3, double Coeff2_3, double Coeff3_3, float freq_central, float bw);
 
 
       virtual void set_order(int order) = 0;
